Consolidate AsyncLazy tests into a single file
Not sure why we have two files at different layers, but it's not needed.
